初步修复php artisan 密令运行

+++ database/migrations/2017_05_08_180345_create_ims_yz_order_address_table.php
@@ -13,22 +13,24 @@ class CreateImsYzOrderAddressTable extends Migration
      */
     public function up()
     {
-        Schema::table('yz_order_address', function (Blueprint $table) {
+        if (Schema::hasTable('yz_order_address')) {
+            Schema::table('yz_order_address', function (Blueprint $table) {
 
-            if (!Schema::hasColumn('yz_order_address', 'province_id')) {
+                if (!Schema::hasColumn('yz_order_address', 'province_id')) {
 
-                $table->integer('province_id')->default(0)->comment('省id');
-            }
+                    $table->integer('province_id')->default(0)->comment('省id');
+                }
 
-            if (!Schema::hasColumn('yz_order_address', 'city_id')) {
+                if (!Schema::hasColumn('yz_order_address', 'city_id')) {
 
-                $table->integer('city_id')->default(0)->comment('市id');
-            }
-            if (!Schema::hasColumn('yz_order_address', 'district_id')) {
+                    $table->integer('city_id')->default(0)->comment('市id');
+                }
+                if (!Schema::hasColumn('yz_order_address', 'district_id')) {
 
-                $table->integer('district_id')->default(0)->comment('区id');
-            }
-        });
+                    $table->integer('district_id')->default(0)->comment('区id');
+                }
+            });
+        }
     }
 
 
@@ -39,11 +41,13 @@ class CreateImsYzOrderAddressTable extends Migration
      */
     public function down()
     {
-        Schema::table('yz_order_address', function (Blueprint $table) {
-            $table->dropColumn('province_id');
-            $table->dropColumn('city_id');
-            $table->dropColumn('district_id');
-        });
+        if (Schema::hasTable('yz_order_address')) {
+            Schema::table('yz_order_address', function (Blueprint $table) {
+                $table->dropColumn('province_id');
+                $table->dropColumn('city_id');
+                $table->dropColumn('district_id');
+            });
+        }
     }
 
 }
