+++ database/migrations/2017_05_18_104602_add_note_to_yz_order_address_table.php
@@ -13,11 +13,13 @@ class AddNoteToYzOrderAddressTable extends Migration
      */
     public function up()
     {
-        Schema::table('yz_order_address', function (Blueprint $table) {
-            if (!Schema::hasColumn('yz_order_address', 'note')) {
-                $table->text('note')->nullable();
-            }
-        });
+        if (Schema::hasTable('yz_order_address')) {
+            Schema::table('yz_order_address', function (Blueprint $table) {
+                if (!Schema::hasColumn('yz_order_address', 'note')) {
+                    $table->text('note')->nullable();
+                }
+            });
+        }
     }
 
     /**
@@ -27,8 +29,10 @@ class AddNoteToYzOrderAddressTable extends Migration
      */
     public function down()
     {
-        Schema::table('yz_order_address', function (Blueprint $table) {
-            $table->dropColumn('note');
-        });
+        if (Schema::hasTable('yz_order_address')) {
+            Schema::table('yz_order_address', function (Blueprint $table) {
+                $table->dropColumn('note');
+            });
+        }
     }
 }
